requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.13
django-cors-headers==4.3.1
djangorestframework==3.15.1
mysqlclient==2.2.4
pkg_resources==0.0.0
sqlparse==0.5.0
typing_extensions==4.12.2
